refactor(modules): extract instance methods for enabling/disabling elements from main module
