+++ src/instanceMethods.js
@@ -1,2 +1,3 @@
 export * from './instanceMethods/hideLoading'
 export * from './instanceMethods/getInput'
+export * from './instanceMethods/enable-disable-elements'
