+++ src/sweetalert2.js
@@ -629,72 +629,6 @@ Object.keys(instanceMethods).forEach(key => {
   }
 })
 
-sweetAlert.enableButtons = () => {
-  if (currentContext) {
-    const {domCache} = currentContext
-    domCache.confirmButton.disabled = false
-    domCache.cancelButton.disabled = false
-  }
-}
-
-sweetAlert.disableButtons = () => {
-  if (currentContext) {
-    const {domCache} = currentContext
-    domCache.confirmButton.disabled = true
-    domCache.cancelButton.disabled = true
-  }
-}
-
-sweetAlert.enableConfirmButton = () => {
-  if (currentContext) {
-    const {domCache} = currentContext
-    domCache.confirmButton.disabled = false
-  }
-}
-
-sweetAlert.disableConfirmButton = () => {
-  if (currentContext) {
-    const {domCache} = currentContext
-    domCache.confirmButton.disabled = true
-  }
-}
-
-sweetAlert.enableInput = () => {
-  if (currentContext) {
-    const input = sweetAlert.getInput()
-    if (!input) {
-      return false
-    }
-    if (input.type === 'radio') {
-      const radiosContainer = input.parentNode.parentNode
-      const radios = radiosContainer.querySelectorAll('input')
-      for (let i = 0; i < radios.length; i++) {
-        radios[i].disabled = false
-      }
-    } else {
-      input.disabled = false
-    }
-  }
-}
-
-sweetAlert.disableInput = () => {
-  if (currentContext) {
-    const input = sweetAlert.getInput()
-    if (!input) {
-      return false
-    }
-    if (input && input.type === 'radio') {
-      const radiosContainer = input.parentNode.parentNode
-      const radios = radiosContainer.querySelectorAll('input')
-      for (let i = 0; i < radios.length; i++) {
-        radios[i].disabled = true
-      }
-    } else {
-      input.disabled = true
-    }
-  }
-}
-
 // Show block with validation error
 sweetAlert.showValidationError = (error) => {
   if (currentContext) {
